Agregados ejercicios del tema 6 y actualizados algunos ejercicios de exámenes

+++ TEMA 5/Ejercicio15.java
@@ -6,11 +6,10 @@ public class Ejercicio15 {
   public static void main(String[] args) {
     System.out.println("Este programa calculará las potencias comprendidas entre la base y el exponente \nde un número introducido.");
     System.out.print("Por favor, introducir una base: ");
-    double base = Double.parseDouble(System.console().readLine());
+    int base = Integer.parseInt(System.console().readLine());
     System.out.print("Por favor, introducir un exponente (entero y positivo): ");
     int exponente = Integer.parseInt(System.console().readLine());
-    double potencia = 0;
-    
+    int potencia = 0;
     if (exponente == 0) {
        System.out.println("Para el número " + base + " exponente " + exponente + " el resultado es: 1");
       } else {
