+++ TEMA 5/Ejercicio21.java
@@ -7,8 +7,8 @@ public class Ejercicio21 {
   public static void main(String[] args) {
     System.out.println("Por favor, introducir números.\nEl programa finalizará cuando se introduzca un número negativo.");
     int cantidad = 0;
-    double impares = 0;
-    double sumaimpares = 0;
+    int impares = 0;
+    int sumaimpares = 0;
     int mayor = 0;
     int num = 0;
     
@@ -29,7 +29,7 @@ public class Ejercicio21 {
       }
       
     System.out.println("La cantidad de números introducidos es: " + cantidad);
-    System.out.printf("La media de los números impares es: %.2f", (sumaimpares / impares));
+    System.out.printf("La media de los números impares es: %.2f", (double)(sumaimpares / impares));
     System.out.println();
     System.out.println("El mayor de los números pares es: " + mayor);
   }
